% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{hasProteinData,EnsDb-method}
\alias{hasProteinData,EnsDb-method}
\alias{hasProteinData}
\title{Determine whether protein data is available in the database}
\usage{
\S4method{hasProteinData}{EnsDb}(x)
}
\arguments{
\item{x}{The \code{\linkS4class{EnsDb}} object.}
}
\value{
A logical of length one, \code{TRUE} if protein annotations are
available and \code{FALSE} otherwise.
}
\description{
Determines whether the \code{\linkS4class{EnsDb}}
provides protein annotation data.
}
\examples{
library(EnsDb.Hsapiens.v75)
## Does this database/package have protein annotations?
hasProteinData(EnsDb.Hsapiens.v75)
}
\seealso{
\code{\link{listTables}}
}
\author{
Johannes Rainer
}
