% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{Deprecated}
\alias{Deprecated}
\alias{ensembldb-deprecated}
\alias{GeneidFilter}
\alias{GenebiotypeFilter}
\alias{EntrezidFilter}
\alias{TxidFilter}
\alias{TxbiotypeFilter}
\alias{ExonidFilter}
\alias{ExonrankFilter}
\alias{SeqnameFilter}
\alias{SeqstrandFilter}
\alias{SeqstartFilter}
\alias{SeqendFilter}
\title{Deprecated functionality}
\usage{
GeneidFilter(value, condition = "==")

GenebiotypeFilter(value, condition = "==")

EntrezidFilter(value, condition = "==")

TxidFilter(value, condition = "==")

TxbiotypeFilter(value, condition = "==")

ExonidFilter(value, condition = "==")

ExonrankFilter(value, condition = "==")

SeqnameFilter(value, condition = "==")

SeqstrandFilter(value, condition = "==")

SeqstartFilter(value, condition = ">", feature = "gene")

SeqendFilter(value, condition = "<", feature = "gene")
}
\arguments{
\item{value}{The value for the filter.}

\item{condition}{The condition for the filter.}

\item{feature}{For \code{SeqstartFilter} and \code{SeqendFilter}: on what type
of feature should the filter be applied? Supported are \code{"gene"},
\code{"tx"} and \code{"exon"}.}
}
\description{
All functions, methods and classes listed on this page are
deprecated and might be removed in future releases.

\code{GeneidFilter} creates a \code{GeneIdFilter}. Use
\code{GeneIdFilter} from the \code{AnnotationFilter} package instead.

\code{GenebiotypeFilter} creates a \code{GeneBiotypeFilter}. Use
\code{GeneBiotypeFilter} from the \code{AnnotationFilter} package instead.

\code{EntrezidFilter} creates a \code{EntrezFilter}. Use
\code{EntrezFilter} from the \code{AnnotationFilter} package instead.

\code{TxidFilter} creates a \code{TxIdFilter}. Use
\code{TxIdFilter} from the \code{AnnotationFilter} package instead.

\code{TxbiotypeFilter} creates a \code{TxBiotypeFilter}. Use
\code{TxBiotypeFilter} from the \code{AnnotationFilter} package instead.

\code{ExonidFilter} creates a \code{ExonIdFilter}. Use
\code{ExonIdFilter} from the \code{AnnotationFilter} package instead.

\code{ExonrankFilter} creates a \code{ExonRankFilter}. Use
\code{ExonRankFilter} from the \code{AnnotationFilter} package instead.

\code{SeqNameFilter} creates a \code{SeqNameFilter}. Use
\code{SeqNameFilter} from the \code{AnnotationFilter} package instead.

\code{SeqstrandFilter} creates a \code{SeqStrandFilter}. Use
\code{SeqStrandFilter} from the \code{AnnotationFilter} instead.

\code{SeqstartFilter} creates a \code{GeneStartFilter}, \code{TxStartFilter}
or \code{ExonStartFilter} depending on the value of the parameter
\code{feature}. Use \code{GeneStartFilter}, \code{TxStartFilter} and
\code{ExonStartFilter} instead.

\code{SeqendFilter} creates a \code{GeneEndFilter}, \code{TxEndFilter}
or \code{ExonEndFilter} depending on the value of the parameter
\code{feature}. Use \code{GeneEndFilter}, \code{TxEndFilter} and
\code{ExonEndFilter} instead.
}
