% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{Deprecated}
\alias{Deprecated}
\alias{EntrezidFilter}
\alias{ExonidFilter}
\alias{ExonrankFilter}
\alias{GenebiotypeFilter}
\alias{GeneidFilter}
\alias{SeqendFilter}
\alias{SeqnameFilter}
\alias{SeqstartFilter}
\alias{SeqstrandFilter}
\alias{TxbiotypeFilter}
\alias{TxidFilter}
\alias{ensembldb-deprecated}
\title{Deprecated functionality}
\usage{
GeneidFilter(value, condition = "==")

GenebiotypeFilter(value, condition = "==")

EntrezidFilter(value, condition = "==")

TxidFilter(value, condition = "==")

TxbiotypeFilter(value, condition = "==")

ExonidFilter(value, condition = "==")

ExonrankFilter(value, condition = "==")

SeqnameFilter(value, condition = "==")

SeqstrandFilter(value, condition = "==")

SeqstartFilter(value, condition = ">", feature = "gene")

SeqendFilter(value, condition = "<", feature = "gene")
}
\arguments{
\item{value}{The value for the filter.}

\item{condition}{The condition for the filter.}

\item{feature}{For \code{SeqstartFilter} and \code{SeqendFilter}: on what type
of feature should the filter be applied? Supported are \code{"gene"},
\code{"tx"} and \code{"exon"}.}
}
\description{
All functions, methods and classes listed on this page are
deprecated and might be removed in future releases.

\code{GeneidFilter} creates a \code{GeneIdFilter}. Use
\code{\link[AnnotationFilters]{GeneIdFilter}} instead.

\code{GenebiotypeFilter} creates a \code{GeneBiotypeFilter}. Use
\code{\link[AnnotationFilters]{GeneBiotypeFilter}} instead.

\code{EntrezidFilter} creates a \code{EntrezFilter}. Use
\code{\link[AnnotationFilters]{EntrezFilter}} instead.

\code{TxidFilter} creates a \code{TxIdFilter}. Use
\code{\link[AnnotationFilters]{TxIdFilter}} instead.

\code{TxbiotypeFilter} creates a \code{TxBiotypeFilter}. Use
\code{\link[AnnotationFilters]{TxBiotypeFilter}} instead.

\code{ExonidFilter} creates a \code{ExonIdFilter}. Use
\code{\link[AnnotationFilters]{ExonIdFilter}} instead.

\code{ExonrankFilter} creates a \code{ExonRankFilter}. Use
\code{\link[AnnotationFilters]{ExonRankFilter}} instead.

\code{SeqNameFilter} creates a \code{SeqNameFilter}. Use
\code{\link[AnnotationFilters]{SeqNameFilter}} instead.

\code{SeqstrandFilter} creates a \code{SeqStrandFilter}. Use
\code{\link[AnnotationFilters]{SeqStrandFilter}} instead.

\code{SeqstartFilter} creates a \code{GeneStartFilter},
\code{TxStartFilter} or \code{ExonStartFilter} depending on the value of the
parameter \code{feature}. Use \code{\link[AnnotationFilters]{GeneStartFilter}},
\code{\link[AnnotationFilters]{TxStartFilter}} and
\code{\link[AnnotationFilters]{ExonStartFilter}} instead.

\code{SeqendFilter} creates a \code{GeneEndFilter},
\code{TxEndFilter} or \code{ExonEndFilter} depending on the value of the
parameter \code{feature}. Use \code{\link[AnnotationFilters]{GeneEndFilter}},
\code{\link[AnnotationFilters]{TxEndFilter}} and
\code{\link[AnnotationFilters]{ExonEndFilter}} instead.
}

