\name{makeEnsembldbPackage}
\alias{ensDbFromGRanges}
\alias{ensDbFromGtf}
\alias{makeEnsembldbPackage}
\alias{fetchTablesFromEnsembl}
\alias{makeEnsemblSQLiteFromTables}
\title{
  Generating a Ensembl annotation package from Ensembl
}
\description{
  These functions allow to retrieve annotations from the Ensembl
  database (\code{fetchTablesFromEnsembl}) create an SQLite database
  from these (\code{makeEnsemblSQLiteFromTables}) and to generate an
  annotation package providing access to this resource
  (\code{makeEnsembldbPackage}).
}
\usage{

ensDbFromGRanges(x, outfile, path, organism, genomeVersion,
                 version, verbose=FALSE)

ensDbFromGtf(gtf, outfile, path, organism, genomeVersion,
             version, verbose=FALSE)

fetchTablesFromEnsembl(version, ensemblapi, user="anonymous",
                       host="ensembldb.ensembl.org", pass="",
                       port=5306, species="human")

makeEnsemblSQLiteFromTables(path=".", dbname)

makeEnsembldbPackage(ensdb, version, maintainer, author,
                     destDir=".", license="Artistic-2.0")

}
\arguments{
  (in alphabetical order)

  \item{ author }{
    The author of the package.
  }

  \item{dbname}{
    The name for the database (optional). By default a name based on the
    species and Ensembl version will be automatically generated (and
    returned by the function).
  }

  \item{ destDir }{
    Where the package should be saved to.
  }

  \item{ ensdb }{
    The file name of the SQLite database generated by \code{makeEnsemblSQLiteFromTables}.
  }

  \item{ ensemblapi }{
    The path to the Ensembl perl API installed locally on the
  system. The Ensembl perl API version has to fit the version.
  }

  \item{genomeVersion}{
    For \code{ensDbFromGtf}: the version of the genome
    (e.g. \code{"GRCh37"}). If not provided the function will try
    to guess it from the file name (assuming file name convention of
    Ensembl GTF files).
  }

  \item{gtf}{
    The GTF file name.
  }

  \item{ host }{
    The hostname to access the Ensembl database.
  }

  \item{ license }{
    The license of the package.
  }

  \item{ maintainer }{
    The maintainer of the package.
  }

  \item{organism}{
    For \code{ensDbFromGtf}: the organism name
    (e.g. \code{"Homo_sapiens"}). If not provided the function will try
    to guess it from the file name (assuming file name convention of
    Ensembl GTF files).
  }

  \item{outfile}{
    The desired file name of the SQLite file. If not provided the name
    of the GTF file will be used.
  }

  \item{ pass }{
    The password for the Ensembl database.
  }

  \item{ path }{
    The directory in which the tables retrieved by
    \code{fetchTablesFromEnsembl} or the SQLite database file generated
    by \code{ensDbFromGtf} are stored.
  }

  \item{ port }{
    The port to be used to connect to the Ensembl database.
  }

  \item{ species }{
    The species for which the annotations should be retrieved.
  }

  \item{ user }{
    The username for the Ensembl database.
  }

  \item{verbose}{
    If progress messages should be shown.
  }

  \item{ version }{
    For \code{fetchTablesFromEnsembl}, \code{ensDbFromGRanges} and \code{ensDbFromGtf}: the
    Ensembl version for which the annotation should be retrieved
    (e.g. 75). The \code{ensDbFromGtf} function will try to guess the
    Ensembl version from the GTF file name if not provided.

    For \code{makeEnsemblDbPackage}: the version for the package.
  }

  \item{x}{
    For \code{ensDbFromGRanges}: the \code{GRanges} object.
  }

}
\details{
  The \code{fetchTablesFromEnsembl} function internally calls the perl
  script \code{get_gene_transcript_exon_tables.pl} to retrieve all
  required information from the Ensembl database using the Ensembl perl
  API.

  As an alternative way, a EnsDb database file can be generated by the
  \code{ensDbFromGtf} from a GTF file from Ensembl or with the
  \code{ensDbFromGRanges} from a \code{GRanges} object e.g. retrieved
  from the AnnotationHub package. The returned database file name can then
  be used as an input to the \code{makeEnsembldbPackage}.
}
\note{
  A local installation of the Ensembl perl API is required for the
  \code{fetchTablesFromEnsembl}. See
  \url{http://www.ensembl.org/info/docs/api/api_installation.html} for
  installation inscructions.

  A database generated from a GTF file lacks some features as they are
  not available in the GTF files from Ensembl. These are: chromosome
  lengths, NCBI Entrezgene IDs.

  \code{GRanges} objects provided by the AnnotationHub package on the
  other hand contain already the required \code{seqinfo} information
  (chromosome length etc) to build an \code{EnsDb} database using the
  \code{ensDbFromGRanges}.

}
\value{
  \code{makeEnsemblSQLiteFromTables}, \code{ensDbFromGRanges} and \code{ensDbFromGtf}: the name
  of the SQLite file.
}
\seealso{
  \code{\link{EnsDb}}, \code{\link{genes}}
}
\author{
Johannes Rainer
}
\examples{

\dontrun{

## get all human gene/transcript/exon annotations from Ensembl (75)
## the resulting tables will be stored by default to the current working
## directory; if the correct Ensembl api (version 75) is defined in the
## PERL5LIB environment variable, the ensemblapi parameter can also be omitted.
fetchTablesFromEnsembl(75,
                       ensemblapi="/home/bioinfo/ensembl/75/API/ensembl/modules",
                       species="human")

## These tables can then be processed to generate a SQLite database
## containing the annotations
DBFile <- makeEnsemblSQLiteFromTables()

## and finally we can generate the package
makeEnsembldbPackage(ensdb=DBFile, version="0.0.1",
                     maintainer="Johannes Rainer <johannes.rainer@eurac.edu>",
                     author="J Rainer")

## Build an annotation file from a GTF file.
## the GTF file can be downloaded from
## ftp://ftp.ensembl.org/pub/release-75/gtf/homo_sapiens/
gtffile <- "Homo_sapiens.GRCh37.75.gtf.gz"
## generate the SQLite database file
DB <- ensDbFromGtf(gtf=paste0(ensemblhost, gtffile), verbose=TRUE)

## load the DB file directly
EDB <- EnsDb(DB)

}

## Generate a sqlite database for genes encoded on chromosome Y
chrY <- system.file("chrY", package="ensembldb")
DBFile <- makeEnsemblSQLiteFromTables(path=chrY ,dbname=tempfile())
## load this database:
edb <- EnsDb(DBFile)

edb

## Generate a sqlite database from a GRanges object specifying
## genes encoded on chromosome Y
load(system.file("YGRanges.RData", package="ensembldb"))

Y

DB <- ensDbFromGRanges(Y, path=tempdir(), version=75,
                       organism="Homo_sapiens")
edb <- EnsDb(DB)


}
\keyword{ data }

